chore(*): migrate tests to run locally
